Refactor EpubChapter to use HtmlChapterParserSingle
This commit refactors the EpubChapter class to use the HtmlChapterParserSingle parser for parsing HTML chapters. The `from_html` method now takes an additional argument `parser` of type `HtmlChapterParserSingle`. The `update_soup` method has been updated to use the `HtmlChapterParserSingle.parse` method to parse the HTML and extract sections. This change improves the code structure and makes it easier to switch between different HTML parsers in the future.